Split jelobjects from jelenvironment tests
